+++ include/editor/editor.h
@@ -297,14 +297,14 @@ extern	int med_save_mine(char *name);
 //	Returns:
 //	 0 = successfully loaded.
 //	 1 = unable to load.
-extern	int med_load_group( char *filename, short *vertex_ids, short *segment_ids, int *num_vertices, int *num_segments);
+extern	int med_load_group( char *filename, int *vertex_ids, short *segment_ids, int *num_vertices, int *num_segments);
 
 // Saves group *filename from disk.
 //	Saves group defined by vertex_ids and segment_ids to disk. 
 //	Returns:
 //	 0 = successfully saved.
 //	 1 = unable to save.
-extern	int med_save_group( char *filename, short *vertex_ids, short *segment_ids, int num_vertices, int num_segments);
+extern	int med_save_group( char *filename, int *vertex_ids, short *segment_ids, int num_vertices, int num_segments);
 
 // Updates the screen... (I put the prototype here for curves.c)
 extern   int medlisp_update_screen();
@@ -434,7 +434,7 @@ extern	void med_assign_uvs_to_side(segment *con_seg, int con_common_side, segmen
 extern	void show_matrix(char *s,vms_matrix *mp,int type);
 
 //	Create coordinate axes in orientation of specified segment, stores vertices at *vp.
-extern	void create_coordinate_axes_from_segment(segment *sp,short *vertnums);
+extern	void create_coordinate_axes_from_segment(segment *sp,int *vertnums);
 
 //	Scale a segment.  Then, if it is connected to something, rotate it.
 extern	int med_scale_and_rotate_segment(segment *seg, vms_angvec *rot);
