+++ main/render.h
@@ -89,10 +89,10 @@ extern int Render_only_bottom;
 void render_start_frame(void);
 
 //Given a lit of point numbers, rotate any that haven't been rotated this frame
-g3s_codes rotate_list(int nv,short *pointnumlist);
+g3s_codes rotate_list(int nv,int *pointnumlist);
 
 //Given a lit of point numbers, project any that haven't been projected
-void project_list(int nv,short *pointnumlist);
+void project_list(int nv,int *pointnumlist);
 
 
 #endif
